+++ apps/cowswap-frontend/src/common/pure/CurrencyArrowSeparator/styled.tsx
@@ -8,11 +8,9 @@ import { loadingAnimationMixin } from './style-mixins'
 export const Box = styled.div<{
   isCollapsed: boolean
   hasSeparatorLine?: boolean
-  disabled: boolean
 }>`
   display: block;
   margin: ${({ isCollapsed }) => (isCollapsed ? '-13px auto' : '2px auto')};
-  cursor: ${({ disabled }) => (disabled ? 'inherit' : 'pointer')};
   color: inherit;
   position: relative;
   z-index: 2;
@@ -20,6 +18,7 @@ export const Box = styled.div<{
   height: 26px;
   justify-content: center;
   transition: width var(${UI.ANIMATION_DURATION}) ease-in-out;
+  pointer-events: none;
 
   ${({ hasSeparatorLine }) =>
     hasSeparatorLine &&
@@ -36,8 +35,9 @@ export const Box = styled.div<{
     `}
 `
 
-export const LoadingWrapper = styled.div<{ isLoading: boolean }>`
+export const LoadingWrapper = styled.button<{ $isLoading: boolean }>`
   --size: 26px;
+
   position: absolute;
   left: calc(50% - var(--size) / 2);
   top: 0;
@@ -54,16 +54,20 @@ export const LoadingWrapper = styled.div<{ isLoading: boolean }>`
   border-radius: 8px;
   width: var(--size);
   margin: auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  pointer-events: auto;
 
-  &:hover {
-    ${({ isLoading }) =>
-      !isLoading &&
-      css`
-        transform: translateY(-2px);
-      `}
+  &:disabled {
+    cursor: not-allowed;
   }
 
-  ${({ isLoading }) => isLoading && loadingAnimationMixin}
+  &:not(:disabled):hover {
+    transform: translateY(-2px);
+  }
+
+  ${({ $isLoading }) => $isLoading && loadingAnimationMixin}
 `
 
 export const ArrowDownIcon = styled(ArrowDown)<{ disabled: boolean }>`
@@ -74,6 +78,6 @@ export const ArrowDownIcon = styled(ArrowDown)<{ disabled: boolean }>`
   padding: 0;
   height: 100%;
   width: 20px;
-  cursor: ${({ disabled }) => (disabled ? 'inherit' : 'pointer')};
+  cursor: inherit;
   color: inherit;
 `