+++ apps/cowswap-frontend/src/modules/application/containers/App/menuConsts.utils.ts
@@ -1,0 +1,6 @@
+import { getExplorerBaseUrl } from '@cowprotocol/common-utils'
+import { SupportedChainId } from '@cowprotocol/cow-sdk'
+
+export function getSolversExplorerUrl(): string {
+  return `${getExplorerBaseUrl(SupportedChainId.MAINNET)}/solvers`
+}