+++ apps/cowswap-frontend/src/modules/ordersTable/pure/ReceiptModal/ReceiptModal.modal.tsx
@@ -57,7 +57,7 @@ interface ReceiptProps {
   limitPrice: Fraction | null
   executionPrice: Fraction | null
   estimatedExecutionPrice: Fraction | null
-  alternativeOrderModalContext: AlternativeOrderModalContext
+  alternativeOrderModalContext?: AlternativeOrderModalContext
 }
 
 const TOOLTIPS_MSG: Record<string, MessageDescriptor> = {